apps/booking/requirements.txt
fastapi
uvicorn
redis
pika
python-jose
PyJWT
starlette>=0.27.0
